Added native project. AudioPlayer interface; Added 6 string sounds; Post-build copying native library;

+++ VoiceChangerApp/ViewModels/WaveletTransformNumericalViewModel.cs
@@ -91,8 +91,8 @@ namespace VoiceChangerApp.ViewModels
             set { SetProperty(ref _points, value); }
         }
 
-        private int _cyclesCount;
-        public int CyclesCount
+        private float _cyclesCount;
+        public float CyclesCount
         {
             get { return _cyclesCount; }
             set { SetProperty(ref _cyclesCount, value); }
